write support for API level 23
API 23 (Android 6.0) throws class cast exceptions when casting
AndroidKeyStoreECPublicKey to RSAPublicKey and will not encrypt data.

This fix ensures that API 23 is now supported.

Updates encryption scheme mode to NONE as RSA does not use a mode (see:
http://crypto.stackexchange.com/questions/25899/using-ecb-as-rsa-encryption-mode-but-encrypted-messages-are-unique)
Finally, change to BouncyCastle (SpongyCastle) instead of using the
built in AndroidOpenSSL implementation as the provider.

[Issue fix: 4]